MODINV-183: Implement action handler for Create Holding action. 1. Classes moved to other packages.
