+++ src/main/java/org/folio/inventory/dataimport/handlers/action/CreateHoldingEventHandler.java
@@ -1,4 +1,9 @@
-package org.folio.inventory.handlers;
+package org.folio.inventory.dataimport.handlers.action;
+
+import static org.apache.commons.lang3.StringUtils.isEmpty;
+import static org.folio.ActionProfile.FolioRecord.HOLDINGS;
+import static org.folio.ActionProfile.FolioRecord.INSTANCE;
+import static org.folio.rest.jaxrs.model.ProfileSnapshotWrapper.ContentType.ACTION_PROFILE;
 
 import java.io.IOException;
 import java.util.Set;
@@ -25,11 +30,6 @@ import io.vertx.core.json.Json;
 import io.vertx.core.json.JsonObject;
 import io.vertx.core.logging.Logger;
 import io.vertx.core.logging.LoggerFactory;
-
-import static org.apache.commons.lang3.StringUtils.isEmpty;
-import static org.folio.ActionProfile.FolioRecord.HOLDINGS;
-import static org.folio.ActionProfile.FolioRecord.INSTANCE;
-import static org.folio.rest.jaxrs.model.ProfileSnapshotWrapper.ContentType.ACTION_PROFILE;
 
 public class CreateHoldingEventHandler implements EventHandler {
